make Bead class a generic class and make a specific class for beads that are used to generate PD codes
